refactor: group runtime/gateway/logging/status into runtime/ module
Move runtime.rs, gateway.rs, logging.rs, status.rs into a single
runtime/ directory module. These files share a startup/lifecycle
responsibility theme.
